refactor class view holder comic

+++ app/src/main/java/com/marvel/heroes/domain/data/dto/Comics.java
@@ -16,7 +16,7 @@ public class Comics {
 
     public String title;
 
-    public double issueNumber;
+    public int issueNumber;
 
     public String variantDescription;
